[REFACTOR] gridview 구조 변경
