+++ js/main/gridView/feature/handleData.js
@@ -1,22 +1,14 @@
-// 랜덤 그리드 && 버튼
-import { clickSubscribe } from "../../utils/clickSubscribe.js";
-import { fetchData } from "../../utils/fetchData.js";
-import { checkPressInLocal } from "../../utils/checkPressInLocal.js";
-
-function renderGridView() {
-  fetchData(".././assets/news.json").then((press) => makeGridView(press));
-}
+import { clickSubscribe } from "../../../utils/clickSubscribe.js";
+import {
+  MIN_PAGE,
+  MAX_PAGE,
+  main_list_page,
+  setListPage,
+} from "./handlePage.js";
+import { handleMouseOver, handleMouseOut } from "./handleEvent.js";
 
 const GRID_NUM = 24;
-const MIN_PAGE = 1;
-const MAX_PAGE = 4;
 const PRESS_NUM = 96;
-
-let main_list_page = MIN_PAGE;
-
-const main_list_ul = document.querySelector(".grid-view-ul");
-const left_btn = document.getElementById("grid-left-btn");
-const right_btn = document.getElementById("grid-right-btn");
 
 const imgIndex = Array(PRESS_NUM)
   .fill()
@@ -26,9 +18,10 @@ function shuffleImgIndex() {
   return [...imgIndex].sort(() => Math.random() - 0.5);
 }
 
-const shuffledPress = shuffleImgIndex();
-
 function showMainList(press) {
+  const main_list_ul = document.querySelector(".grid-view-ul");
+  const shuffledPress = shuffleImgIndex();
+
   main_list_ul.innerHTML = "";
   for (
     let i = GRID_NUM * (main_list_page - 1);
@@ -49,6 +42,7 @@ function showMainList(press) {
       handleMouseOut(_img, ` ${press[shuffledPress[i]].lightSrc}`)
     );
 
+    //구독하기 해제하기 클릭 시
     _img.addEventListener("click", (e) =>
       clickSubscribe(e.target.parentElement.dataset.press)
     );
@@ -58,29 +52,20 @@ function showMainList(press) {
   }
 }
 
-function handleMouseOver(_img, press) {
-  if (checkPressInLocal(press)) {
-    _img.setAttribute("src", `../images/icon/Unsubscribe.svg`);
-  } else {
-    _img.setAttribute("src", "../images/icon/Subscribe.svg");
-  }
-}
-
-function handleMouseOut(_img, originImg) {
-  _img.setAttribute("src", originImg);
-}
-
 function changePage(e, press) {
   if (e.target.id === "grid-left") {
-    main_list_page--;
+    setListPage(main_list_page - 1);
   } else {
-    main_list_page++;
+    setListPage(main_list_page + 1);
   }
   showMainList(press);
   checkPage();
 }
 
 function checkPage() {
+  const left_btn = document.getElementById("grid-left-btn");
+  const right_btn = document.getElementById("grid-right-btn");
+
   if (main_list_page === MIN_PAGE) left_btn.style.visibility = "hidden";
   else if (main_list_page === MAX_PAGE) right_btn.style.visibility = "hidden";
   else {
@@ -89,11 +74,4 @@ function checkPage() {
   }
 }
 
-function makeGridView(press) {
-  right_btn.addEventListener("click", (e) => changePage(e, press));
-  left_btn.addEventListener("click", (e) => changePage(e, press));
-  showMainList(press);
-  checkPage();
-}
-
-export { renderGridView, main_list_page, MIN_PAGE, MAX_PAGE };
+export { changePage, showMainList, checkPage };
