+++ js/main/gridView/feature/handleEvent.js
@@ -1,0 +1,15 @@
+import { checkPressInLocal } from "../../../utils/checkPressInLocal.js";
+
+function handleMouseOver(_img, press) {
+  if (checkPressInLocal(press)) {
+    _img.setAttribute("src", `../images/icon/Unsubscribe.svg`);
+  } else {
+    _img.setAttribute("src", "../images/icon/Subscribe.svg");
+  }
+}
+
+function handleMouseOut(_img, originImg) {
+  _img.setAttribute("src", originImg);
+}
+
+export { handleMouseOver, handleMouseOut };
